requirements.txt
pillow==10.3.0
termcolor==2.4.0